Blunt move and sed rename
